i cant decide what im doing tbh
i think im gonna use redux toolkit without the whole query setup because query makes it too complicated for what im trying to do and i'd rather have personal control over it all tbh
