+++ app/lib/redux/rootReducer.ts
@@ -1,6 +1,6 @@
 /* Instruments */
-import { counterSlice } from './slices'
+import { systemSlice } from './slices'
 
 export const reducer = {
-  counter: counterSlice.reducer,
+  system: systemSlice.reducer,
 }
