+++ app/lib/redux/slices/index.ts
@@ -1,1 +1,1 @@
-export * from './schedulesSlice'
+export * from './systemSlice'
