+++ app/lib/redux/slices/systemSlice/index.ts
@@ -1,3 +1,3 @@
-export * from './schedulesSlice'
+export * from './systemSlice'
 export * from './thunks'
 export * from './selectors'
