+++ app/lib/redux/slices/systemSlice/systemSlice.ts
@@ -1,0 +1,78 @@
+/* Core */
+import { createSlice, type PayloadAction } from '@reduxjs/toolkit'
+import { useSession } from 'next-auth/react'
+import { staticGenerationAsyncStorage } from 'next/dist/client/components/static-generation-async-storage'
+
+/* Instruments */
+import { loadData, updateSchedules } from './thunks'
+
+const initialState: SystemSliceState = {
+  schedules: [],
+  currentScheduleIndex: -1,
+  status: 'idle',
+}
+
+export const systemSlice = createSlice({
+  name: 'system',
+  initialState,
+  // The `reducers` field lets us define reducers and generate associated actions
+  reducers: {
+    addClass: (state, action: PayloadAction<number>) => {
+      // Redux Toolkit allows us to write "mutating" logic in reducers. It
+      // doesn't actually mutate the state because it uses the Immer library,
+      // which detects changes to a "draft state" and produces a brand new
+      // immutable state based off those changes
+      state.schedules[state.currentScheduleIndex].class_numbers.push(action.payload)
+    },
+    removeClass: (state, action: PayloadAction<number>) => {
+      state.schedules[state.currentScheduleIndex].class_numbers = state.schedules[state.currentScheduleIndex].class_numbers.filter(obj => obj != action.payload);
+    },
+    // Use the PayloadAction type to declare the contents of `action.payload`
+    selectSchedule: (state, action: PayloadAction<number>) => {
+      if (state.schedules.length > action.payload  && action.payload > -1) {
+        state.currentScheduleIndex = action.payload;
+      }
+    },
+  },
+  // The `extraReducers` field lets the slice handle actions defined elsewhere,
+  // including actions generated by createAsyncThunk or in other slices.
+  extraReducers: (builder) => {
+    builder
+      .addCase(updateSchedules.pending, (state) => {
+        state.status = 'loading'
+      })
+      .addCase(updateSchedules.fulfilled, (state, action) => {
+        state.status = 'idle'
+        state.schedules = action.payload;
+      })
+      .addCase(loadData.fulfilled, (state, action) =>
+      {
+        const { data: session, status } = useSession()
+        state.status = status === "authenticated" ? 'cached' : 'idle';
+
+        if (action.payload?.schedules && action.payload.currentScheduleIndex) {
+          state.schedules = action.payload.schedules;
+          state.currentScheduleIndex = action.payload.currentScheduleIndex;
+        }
+      })
+  },
+})
+
+export interface Schedule {
+  class_numbers: Array<number>;
+  term: string;
+  name: string;
+  id: string;
+}
+
+/* Types */
+export interface SystemSliceState {
+  schedules: Array<Schedule>
+  currentScheduleIndex: number
+  status: 'idle' | 'loading' | 'cached' | 'failed'
+}
+
+export interface SystemData {
+  schedules: Array<Schedule>
+  currentScheduleIndex: number
+}
